requirements.txt
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-chroma
chromadb
pyyaml
sqlparse
python-dotenv
colorama
rich